Variable size check in Capability programming.
[Issue Description]
Variable size checking tool reports an error of overflow
possible on NextCap variable.

[Resolution]
Change of variable size.

Package/Module:
OneSiliconPkg

[Impacted Platform]
ALL.

Hsd-es-id: 18044220894
Original commit date: Thu Feb 19 16:51:47 2026 +0100
Original commit hash: 3082116515d6e7491fb0a8389dae9ba8fe7a1205

+++ OneSiliconPkg/IpBlock/TcPcieRp/IpTcPcieRpLib/IpTcPcieRpLib.c
@@ -70,7 +70,7 @@ IpTcPcieInitCapabilityList (
   UINT32      Data32Or;
   UINT16      Data16;
   UINT8       Data8;
-  UINT16      NextCap;
+  UINT32      NextCap;
 
   ///
   /// If not valid root port configuration space then return
